Organize components and containers
