+++ src/containers/FavoritePoster/FavoritePoster.test.js
