+++ src/containers/Favorites/Favorites.test.js
