+++ src/containers/MoviePoster/MoviePoster.js
