+++ src/containers/MoviePoster/MoviePoster.test.js
